README.md
MEXC referral code mexc-MXUSDT is the best way for new users to begin trading with maximum value, thanks to the bonus up to 8000 USDT and up to 50% off fees. With simple activation steps, a powerful platform, and numerous trading opportunities, MEXC delivers an excellent experience for newcomers and professionals alike.